+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
-All Notable changes to `sancus` will be documented in this file
+All Notable changes to `sancus` will be documented in this file.
+
+## 2016-09-13
+
+- Renamed `score()` function of `WilsonInterval` class as `getScore()`
+- Updated tests
+- Updated docs
 
 ## 2015-10-19
 
-### Added
 - Wilson Interval
 - Wald Interval
 - Agresti-Coull Interval